requirements.txt
Flask==3.1.0
flask-restx==1.3.0
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
boto3